requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
Django==4.2.5
django-bootstrap-v5==1.0.11
django-bootstrap5==23.3
django-filter==23.2
djangorestframework==3.14.0
psycopg==3.1.12
psycopg-binary==3.1.12
pytz==2023.3
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.7.1
uWSGI==2.0.22
